Add AppBar class widget to reuse in other screen ,
