+++ open_sism/lib/models/constants.dart
@@ -3,3 +3,15 @@ import 'package:flutter/material.dart';
 const kPrimaryColor = Color(0xFF84AB5C);
 const kTextColor = Color(0xFF202E2E);
 const kTextLigntColor = Color(0xFF7286A5);
+
+const kAppBarHeight = Size.fromHeight(50);
+const kAppBarBackgroundColor = Color(0xff005cb2);
+const kAppBarLeadingIconColor = Color(0xff002984);
+const kAppBarTitleColor = Colors.white60;
+const kAppBarTextStyle = TextStyle(
+  letterSpacing: 2,
+  fontSize: 30,
+  color: kAppBarTitleColor,
+  //fontFamily: ,
+  fontWeight: FontWeight.bold,
+);
